Conditional statements contribution to Operator
